+++ rust-k256/src/lib.rs
@@ -7,14 +7,16 @@ use k256::{
     elliptic_curve::sec1::ToEncodedPoint,
     elliptic_curve::{bigint::ArrayEncoding, group::ff::PrimeField},
     sha2::{digest::Output, Digest, Sha256},
-    FieldBytes, ProjectivePoint, Scalar, Secp256k1, U256,
+    FieldBytes, Scalar, Secp256k1, U256,
 }; // requires 'getrandom' feature
 use std::panic;
+// TODO #86
+pub use k256::ProjectivePoint;
 
 const L: usize = 48;
 const COUNT: usize = 2;
 const OUT: usize = L * COUNT;
-const DST: &[u8] = b"QUUX-V01-CS02-with-secp256k1_XMD:SHA-256_SSWU_RO_"; // Hash to curve algorithm
+pub const DST: &[u8] = b"QUUX-V01-CS02-with-secp256k1_XMD:SHA-256_SSWU_RO_"; // Hash to curve algorithm
 
 fn print_type_of<T>(_: &T) {
     println!("{}", std::any::type_name::<T>());
@@ -159,272 +161,28 @@ fn byte_array_to_scalar(bytes: &[u8]) -> Scalar {
 
 #[cfg(test)]
 mod tests {
-    /* Module consists of two tests; one for each type of signature. One of them also do printings of the values, which can be useful to you when
-    comparing different implementations (the values are also asserted by that one). 
-    Their setup is shared, `mod helpers` contains not refactored code, which is still instrumental to the tests.*/
-
     use super::*;
-    use helpers::{gen_test_scalar_sk, hash_to_secp, test_gen_signals, PlumeVersion};
+    use hex_literal::hex;
 
-    const G: ProjectivePoint = ProjectivePoint::GENERATOR;
-    const M: &[u8; 29] = b"An example app message string";
-
-    fn test_signature<'a>(version: PlumeVersion) -> PlumeSignature<'a> {
-        // Fixed key nullifier, secret key, and random value for testing
-        // Normally a secure enclave would generate these values, and output to a wallet implementation
-        let (
-            pk, nullifier, c, r_sk_c, g_r, 
-            hash_m_pk_pow_r
-        ) = test_gen_signals(M, version);
-
-        // The signer's secret key. It is only accessed within the secure enclave.
-        let sk = gen_test_scalar_sk();
-
-        // The user's public key: g^sk.
-        let pk = &G * &sk;
-
-        // Test encode_pt()
-        let g_as_bytes = encode_pt(&G);
+    // Test encode_pt()
+    #[test]
+    fn test_encode_pt() {
+        let g_as_bytes = encode_pt(&ProjectivePoint::GENERATOR);
         assert_eq!(
             hex::encode(g_as_bytes),
             "0279be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798"
         );
-        // Test byte_array_to_scalar()
-        let scalar = byte_array_to_scalar(&c); // TODO this `fn` looks suspicious as in reproducing const time ops
+    }
+
+    // Test byte_array_to_scalar()
+    #[test]
+    fn test_byte_array_to_scalar() {
+        let scalar = byte_array_to_scalar(&hex!(
+            "c6a7fc2c926ddbaf20731a479fb6566f2daa5514baae5223fe3b32edbce83254"
+        )); // TODO this `fn` looks suspicious as in reproducing const time ops
         assert_eq!(
             hex::encode(scalar.to_bytes()),
             "c6a7fc2c926ddbaf20731a479fb6566f2daa5514baae5223fe3b32edbce83254"
         );
-        // Test the hash-to-curve algorithm
-        let h = hash_to_secp(b"abc");
-        assert_eq!(
-            hex::encode(h.to_affine().to_encoded_point(false).x().unwrap()),
-            "3377e01eab42db296b512293120c6cee72b6ecf9f9205760bd9ff11fb3cb2c4b"
-        );
-        assert_eq!(
-            hex::encode(h.to_affine().to_encoded_point(false).y().unwrap()),
-            "7f95890f33efebd1044d382a01b1bee0900fb6116f94688d487c6c7b9c8371f6"
-        );
-
-        // Print nullifier
-        println!(
-            "nullifier.x: {:?}",
-            hex::encode(nullifier.to_affine().to_encoded_point(false).x().unwrap())
-        );
-        println!(
-            "nullifier.y: {:?}",
-            hex::encode(nullifier.to_affine().to_encoded_point(false).y().unwrap())
-        );
-        // Print c
-        println!("c: {:?}", hex::encode(&c));
-        // Print r_sk_c
-        println!("r_sk_c: {:?}", hex::encode(r_sk_c.to_bytes()));
-        // Print g_r
-        println!(
-            "g_r.x: {:?}",
-            hex::encode(
-                g_r.unwrap()
-                    .to_affine()
-                    .to_encoded_point(false)
-                    .x()
-                    .unwrap()
-            )
-        );
-        println!(
-            "g_r.y: {:?}",
-            hex::encode(
-                g_r.unwrap()
-                    .to_affine()
-                    .to_encoded_point(false)
-                    .y()
-                    .unwrap()
-            )
-        );
-        // Print hash_m_pk_pow_r
-        println!(
-            "hash_m_pk_pow_r.x: {:?}",
-            hex::encode(
-                hash_m_pk_pow_r
-                    .unwrap()
-                    .to_affine()
-                    .to_encoded_point(false)
-                    .x()
-                    .unwrap()
-            )
-        );
-        println!(
-            "hash_m_pk_pow_r.y: {:?}",
-            hex::encode(
-                hash_m_pk_pow_r
-                    .unwrap()
-                    .to_affine()
-                    .to_encoded_point(false)
-                    .y()
-                    .unwrap()
-            )
-        );
-
-        PlumeSignature::<'a> {
-            message: M,
-            pk: &pk,
-            nullifier: &nullifier,
-            c: &c,
-            s: &r_sk_c,
-            v1: match version {
-                PlumeVersion::V1 => Some(PlumeSignatureV1Fields {
-                    r_point: &g_r.unwrap(),
-                    hashed_to_curve_r: &hash_m_pk_pow_r.unwrap(),
-                }),
-                PlumeVersion::V2 => None,
-            },
-        }
-    }
-    
-    // Verify the signals, normally this would happen in ZK with only the nullifier public, which would have a zk verifier instead
-    // The wallet should probably run this prior to snarkify-ing as a sanity check
-    // `M` and nullifier should be public, so we can verify that they are correct
-    
-    #[test]
-    fn plume_v1_test() {
-        let verified = test_signature(PlumeVersion::V1).verify_signals();
-        println!("Verified: {}", verified);
-        assert!(verified);
-    }
-
-    #[test]
-    fn plume_v2_test() {
-        assert!(test_signature(PlumeVersion::V2).verify_signals());
-    }
-
-    mod helpers {
-        use super::*;
-        use hex_literal::hex;
-
-        #[derive(Debug)]
-        pub enum PlumeVersion {
-            V1,
-            V2,
-        }
-
-        // Generates a deterministic secret key for deterministic testing. Should be replaced by random oracle in production deployments.
-        pub fn gen_test_scalar_sk() -> Scalar {
-            Scalar::from_repr(
-                hex!("519b423d715f8b581f4fa8ee59f4771a5b44c8130b4e3eacca54a56dda72b464").into(),
-            )
-            .unwrap()
-        }
-
-        // Generates a deterministic r for deterministic testing. Should be replaced by random oracle in production deployments.
-        fn gen_test_scalar_r() -> Scalar {
-            Scalar::from_repr(
-                hex!("93b9323b629f251b8f3fc2dd11f4672c5544e8230d493eceea98a90bda789808").into(),
-            )
-            .unwrap()
-        }
-
-        // Calls the hash to curve function for secp256k1, and returns the result as a ProjectivePoint
-        pub fn hash_to_secp(s: &[u8]) -> ProjectivePoint {
-            let pt: ProjectivePoint = Secp256k1::hash_from_bytes::<ExpandMsgXmd<Sha256>>(
-                &[s],
-                //b"CURVE_XMD:SHA-256_SSWU_RO_"
-                &[DST],
-            )
-            .unwrap();
-            pt
-        }
-
-        // These generate test signals as if it were passed from a secure enclave to wallet. Note that leaking these signals would leak pk, but not sk.
-        // Outputs these 6 signals, in this order
-        // g^sk																(private)
-        // hash[m, pk]^sk 													public nullifier
-        // c = hash2(g, pk, hash[m, pk], hash[m, pk]^sk, gr, hash[m, pk]^r)	(public or private)
-        // r + sk * c														(public or private)
-        // g^r																(private, optional)
-        // hash[m, pk]^r													(private, optional)
-        pub fn test_gen_signals(
-            m: &[u8],
-            version: PlumeVersion,
-        ) -> (
-            ProjectivePoint,
-            ProjectivePoint,
-            Output<Sha256>,
-            Scalar,
-            Option<ProjectivePoint>,
-            Option<ProjectivePoint>,
-        ) {
-            // The base point or generator of the curve.
-            let g = ProjectivePoint::GENERATOR;
-
-            // The signer's secret key. It is only accessed within the secure enclave.
-            let sk = gen_test_scalar_sk();
-
-            // A random value r. It is only accessed within the secure enclave.
-            let r = gen_test_scalar_r();
-
-            // The user's public key: g^sk.
-            let pk = &g * &sk;
-
-            // The generator exponentiated by r: g^r.
-            let g_r = &g * &r;
-
-            // hash[m, pk]
-            let hash_m_pk = hash_to_curve(m, &pk).unwrap();
-
-            println!(
-                "h.x: {:?}",
-                hex::encode(hash_m_pk.to_affine().to_encoded_point(false).x().unwrap())
-            );
-            println!(
-                "h.y: {:?}",
-                hex::encode(hash_m_pk.to_affine().to_encoded_point(false).y().unwrap())
-            );
-
-            // hash[m, pk]^r
-            let hash_m_pk_pow_r = &hash_m_pk * &r;
-            println!(
-                "hash_m_pk_pow_r.x: {:?}",
-                hex::encode(
-                    hash_m_pk_pow_r
-                        .to_affine()
-                        .to_encoded_point(false)
-                        .x()
-                        .unwrap()
-                )
-            );
-            println!(
-                "hash_m_pk_pow_r.y: {:?}",
-                hex::encode(
-                    hash_m_pk_pow_r
-                        .to_affine()
-                        .to_encoded_point(false)
-                        .y()
-                        .unwrap()
-                )
-            );
-
-            // The public nullifier: hash[m, pk]^sk.
-            let nullifier = &hash_m_pk * &sk;
-
-            // The Fiat-Shamir type step.
-            let c = match version {
-                PlumeVersion::V1 => c_sha256_vec_signal(vec![
-                    &g,
-                    &pk,
-                    &hash_m_pk,
-                    &nullifier,
-                    &g_r,
-                    &hash_m_pk_pow_r,
-                ]),
-                PlumeVersion::V2 => c_sha256_vec_signal(vec![&nullifier, &g_r, &hash_m_pk_pow_r]),
-            };
-            dbg!(&c, version);
-
-            let c_scalar = &Scalar::reduce_nonzero(U256::from_be_byte_array(c.to_owned()));
-            // This value is part of the discrete log equivalence (DLEQ) proof.
-            let r_sk_c = r + sk * c_scalar;
-
-            // Return the signature.
-            (pk, nullifier, c, r_sk_c, Some(g_r), Some(hash_m_pk_pow_r))
-        }
     }
 }
