+++ rust-k256/tests/common/mod.rs
@@ -1,0 +1,5 @@
+/* Module consists of two tests; one for each type of signature. One of them also do printings of the values, which can be useful to you when
+comparing different implementations (the values are also asserted by that one). 
+Their setup is shared, `mod helpers` contains barely not refactored code, which is still instrumental to the tests.*/
+
+use super::*;
